Añadiendo a los campos esperados como no nulos se validen al ser creadas entidades. Cambio de int a long para las variables usadas como PrimaryKey.

+++ init.sql
@@ -18,11 +18,11 @@ USE `film_affinity_fruna` ;
 -- Table `film_affinity_fruna`.`usuario`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `film_affinity_fruna`.`usuario` (
-                                                               `id_usuario` INT NOT NULL,
+                                                               `id_usuario` BIGINT NOT NULL,
                                                                `nombre` VARCHAR(45) NOT NULL,
-                                                               `correo` VARCHAR(45) NOT NULL,
-                                                               `contrasennia` VARCHAR(45) NOT NULL,
-                                                               PRIMARY KEY (`id_usuario`))
+    `correo` VARCHAR(45) NOT NULL,
+    `contrasennia` VARCHAR(45) NOT NULL,
+    PRIMARY KEY (`id_usuario`))
     ENGINE = InnoDB;
 
 
@@ -31,8 +31,8 @@ CREATE TABLE IF NOT EXISTS `film_affinity_fruna`.`usuario` (
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `film_affinity_fruna`.`genero` (
                                                               `nombre` VARCHAR(45) NOT NULL,
-                                                              `descripcion` VARCHAR(45) NULL,
-                                                              PRIMARY KEY (`nombre`))
+    `descripcion` VARCHAR(45) NULL,
+    PRIMARY KEY (`nombre`))
     ENGINE = InnoDB;
 
 
@@ -40,13 +40,13 @@ CREATE TABLE IF NOT EXISTS `film_affinity_fruna`.`genero` (
 -- Table `film_affinity_fruna`.`director`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `film_affinity_fruna`.`director` (
-                                                                `id_director` INT NOT NULL,
+                                                                `id_director` BIGINT NOT NULL,
                                                                 `nombre` VARCHAR(45) NOT NULL,
-                                                                `pais` VARCHAR(45) NOT NULL,
-                                                                `fecha_nacimiento` DATE NOT NULL,
-                                                                `fecha_defuncion` DATE NULL,
-                                                                `enlace_wiki` VARCHAR(80) NULL,
-                                                                PRIMARY KEY (`id_director`))
+    `pais` VARCHAR(45) NOT NULL,
+    `fecha_nacimiento` DATE NOT NULL,
+    `fecha_defuncion` DATE NULL,
+    `enlace_wiki` VARCHAR(80) NULL,
+    PRIMARY KEY (`id_director`))
     ENGINE = InnoDB;
 
 
@@ -54,28 +54,28 @@ CREATE TABLE IF NOT EXISTS `film_affinity_fruna`.`director` (
 -- Table `film_affinity_fruna`.`pelicula`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `film_affinity_fruna`.`pelicula` (
-                                                                `id_pelicula` INT NOT NULL AUTO_INCREMENT,
+                                                                `id_pelicula` BIGINT NOT NULL AUTO_INCREMENT,
                                                                 `nombre` VARCHAR(45) NOT NULL,
-                                                                `pais` VARCHAR(45) NULL,
-                                                                `duracion` TIME NOT NULL,
-                                                                `estreno` DATE NOT NULL,
-                                                                `sinopsis` TEXT NULL,
-                                                                `enlace_wiki` VARCHAR(80) NULL,
-                                                                `genero_nombre` VARCHAR(45) NOT NULL,
-                                                                `director_id_director` INT NOT NULL,
-                                                                PRIMARY KEY (`id_pelicula`),
-                                                                INDEX `fk_pelicula_genero1_idx` (`genero_nombre` ASC) VISIBLE,
-                                                                INDEX `fk_pelicula_director1_idx` (`director_id_director` ASC) VISIBLE,
-                                                                CONSTRAINT `fk_pelicula_genero1`
-                                                                    FOREIGN KEY (`genero_nombre`)
-                                                                        REFERENCES `film_affinity_fruna`.`genero` (`nombre`)
-                                                                        ON DELETE NO ACTION
-                                                                        ON UPDATE NO ACTION,
-                                                                CONSTRAINT `fk_pelicula_director1`
-                                                                    FOREIGN KEY (`director_id_director`)
-                                                                        REFERENCES `film_affinity_fruna`.`director` (`id_director`)
-                                                                        ON DELETE NO ACTION
-                                                                        ON UPDATE NO ACTION)
+    `pais` VARCHAR(45) NULL,
+    `duracion` TIME NOT NULL,
+    `estreno` DATE NOT NULL,
+    `sinopsis` TEXT NULL,
+    `enlace_wiki` VARCHAR(80) NULL,
+    `genero_nombre` VARCHAR(45) NOT NULL,
+    `director_id_director` BIGINT NOT NULL,
+    PRIMARY KEY (`id_pelicula`),
+    INDEX `fk_pelicula_genero1_idx` (`genero_nombre` ASC) VISIBLE,
+    INDEX `fk_pelicula_director1_idx` (`director_id_director` ASC) VISIBLE,
+    CONSTRAINT `fk_pelicula_genero1`
+    FOREIGN KEY (`genero_nombre`)
+    REFERENCES `film_affinity_fruna`.`genero` (`nombre`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION,
+    CONSTRAINT `fk_pelicula_director1`
+    FOREIGN KEY (`director_id_director`)
+    REFERENCES `film_affinity_fruna`.`director` (`id_director`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION)
     ENGINE = InnoDB;
 
 
@@ -83,13 +83,13 @@ CREATE TABLE IF NOT EXISTS `film_affinity_fruna`.`pelicula` (
 -- Table `film_affinity_fruna`.`actor`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `film_affinity_fruna`.`actor` (
-                                                             `id_actor` INT NOT NULL AUTO_INCREMENT,
+                                                             `id_actor` BIGINT NOT NULL AUTO_INCREMENT,
                                                              `nombre` VARCHAR(45) NOT NULL,
-                                                             `pais` VARCHAR(45) NOT NULL,
-                                                             `fecha_nacimiento` DATE NOT NULL,
-                                                             `fecha_defuncion` DATE NULL,
-                                                             `enlace_wiki` VARCHAR(80) NULL,
-                                                             PRIMARY KEY (`id_actor`))
+    `pais` VARCHAR(45) NOT NULL,
+    `fecha_nacimiento` DATE NOT NULL,
+    `fecha_defuncion` DATE NULL,
+    `enlace_wiki` VARCHAR(80) NULL,
+    PRIMARY KEY (`id_actor`))
     ENGINE = InnoDB;
 
 
@@ -97,23 +97,23 @@ CREATE TABLE IF NOT EXISTS `film_affinity_fruna`.`actor` (
 -- Table `film_affinity_fruna`.`reparto`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `film_affinity_fruna`.`reparto` (
-                                                               `id_reparto` INT NOT NULL AUTO_INCREMENT,
+                                                               `id_reparto` BIGINT NOT NULL AUTO_INCREMENT,
                                                                `nombre_personaje` VARCHAR(45) NOT NULL,
-                                                               `actor_id_actor` INT NOT NULL,
-                                                               `pelicula_id_pelicula` INT NOT NULL,
-                                                               PRIMARY KEY (`id_reparto`),
-                                                               INDEX `fk_reparto_actor_idx` (`actor_id_actor` ASC) VISIBLE,
-                                                               INDEX `fk_reparto_pelicula1_idx` (`pelicula_id_pelicula` ASC) VISIBLE,
-                                                               CONSTRAINT `fk_reparto_actor`
-                                                                   FOREIGN KEY (`actor_id_actor`)
-                                                                       REFERENCES `film_affinity_fruna`.`actor` (`id_actor`)
-                                                                       ON DELETE NO ACTION
-                                                                       ON UPDATE NO ACTION,
-                                                               CONSTRAINT `fk_reparto_pelicula1`
-                                                                   FOREIGN KEY (`pelicula_id_pelicula`)
-                                                                       REFERENCES `film_affinity_fruna`.`pelicula` (`id_pelicula`)
-                                                                       ON DELETE NO ACTION
-                                                                       ON UPDATE NO ACTION)
+    `actor_id_actor` BIGINT NOT NULL,
+    `pelicula_id_pelicula` BIGINT NOT NULL,
+    PRIMARY KEY (`id_reparto`),
+    INDEX `fk_reparto_actor_idx` (`actor_id_actor` ASC) VISIBLE,
+    INDEX `fk_reparto_pelicula1_idx` (`pelicula_id_pelicula` ASC) VISIBLE,
+    CONSTRAINT `fk_reparto_actor`
+    FOREIGN KEY (`actor_id_actor`)
+    REFERENCES `film_affinity_fruna`.`actor` (`id_actor`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION,
+    CONSTRAINT `fk_reparto_pelicula1`
+    FOREIGN KEY (`pelicula_id_pelicula`)
+    REFERENCES `film_affinity_fruna`.`pelicula` (`id_pelicula`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION)
     ENGINE = InnoDB;
 
 
@@ -121,21 +121,21 @@ CREATE TABLE IF NOT EXISTS `film_affinity_fruna`.`reparto` (
 -- Table `film_affinity_fruna`.`usuario_has_pelicula`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `film_affinity_fruna`.`usuario_has_pelicula` (
-                                                                            `usuario_id_usuario` INT NOT NULL,
-                                                                            `pelicula_id_pelicula` INT NOT NULL,
+                                                                            `usuario_id_usuario` BIGINT NOT NULL,
+                                                                            `pelicula_id_pelicula` BIGINT NOT NULL,
                                                                             PRIMARY KEY (`usuario_id_usuario`, `pelicula_id_pelicula`),
-                                                                            INDEX `fk_usuario_has_pelicula_pelicula1_idx` (`pelicula_id_pelicula` ASC) VISIBLE,
-                                                                            INDEX `fk_usuario_has_pelicula_usuario1_idx` (`usuario_id_usuario` ASC) VISIBLE,
-                                                                            CONSTRAINT `fk_usuario_has_pelicula_usuario1`
-                                                                                FOREIGN KEY (`usuario_id_usuario`)
-                                                                                    REFERENCES `film_affinity_fruna`.`usuario` (`id_usuario`)
-                                                                                    ON DELETE NO ACTION
-                                                                                    ON UPDATE NO ACTION,
-                                                                            CONSTRAINT `fk_usuario_has_pelicula_pelicula1`
-                                                                                FOREIGN KEY (`pelicula_id_pelicula`)
-                                                                                    REFERENCES `film_affinity_fruna`.`pelicula` (`id_pelicula`)
-                                                                                    ON DELETE NO ACTION
-                                                                                    ON UPDATE NO ACTION)
+    INDEX `fk_usuario_has_pelicula_pelicula1_idx` (`pelicula_id_pelicula` ASC) VISIBLE,
+    INDEX `fk_usuario_has_pelicula_usuario1_idx` (`usuario_id_usuario` ASC) VISIBLE,
+    CONSTRAINT `fk_usuario_has_pelicula_usuario1`
+    FOREIGN KEY (`usuario_id_usuario`)
+    REFERENCES `film_affinity_fruna`.`usuario` (`id_usuario`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION,
+    CONSTRAINT `fk_usuario_has_pelicula_pelicula1`
+    FOREIGN KEY (`pelicula_id_pelicula`)
+    REFERENCES `film_affinity_fruna`.`pelicula` (`id_pelicula`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION)
     ENGINE = InnoDB;
